+++ hw-ledger-transports/node.ts
@@ -1,10 +1,10 @@
 // Copyright 2017-2023 @polkadot/hw-ledger authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type Transport from 'https://esm.sh/@ledgerhq/hw-transport@6.27.9';
+import type Transport from 'https://esm.sh/@ledgerhq/hw-transport@6.27.10';
 import type { TransportDef } from './types.ts';
 
-import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.28.6';
+import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.28.7';
 
 export { packageInfo } from './packageInfo.ts';