+++ util-crypto/test/index.ts
@@ -1,0 +1,6 @@
+// Copyright 2017-2023 @polkadot/util-crypto authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { perf, perfCmp } from 'https://deno.land/x/polkadot/util/test/index.ts';
+
+export { perfWasm } from './performance.ts';